+++ mmocr/models/textdet/modules/proposal_local_graph.py
@@ -5,7 +5,7 @@ import torch
 from lanms import merge_quadrangle_n9 as la_nms
 from mmcv.ops import RoIAlignRotated
 
-from mmocr.models.textdet.postprocess.wrapper import fill_hole
+from mmocr.models.textdet.postprocess.utils import fill_hole
 from .utils import (euclidean_distance_matrix, feature_embedding,
                     normalize_adjacent_matrix)
