+++ mmocr/models/textdet/postprocess/drrg_postprocessor.py
@@ -1,7 +1,7 @@
 from mmocr.models.builder import POSTPROCESSOR
 from .base_postprocessor import BasePostprocessor
-from .wrapper import (clusters2labels, comps2boundaries, connected_components,
-                      graph_propagation, remove_single)
+from .utils import (clusters2labels, comps2boundaries, connected_components,
+                    graph_propagation, remove_single)
 
 
 @POSTPROCESSOR.register_module()
