+++ tests/test_utils/test_wrapper.py
@@ -6,7 +6,7 @@ import torch
 from mmocr.models.textdet.postprocess import (DBPostprocessor,
                                               FCEPostprocessor,
                                               TextSnakePostprocessor)
-from mmocr.models.textdet.postprocess.wrapper import comps2boundaries, poly_nms
+from mmocr.models.textdet.postprocess.utils import comps2boundaries, poly_nms
 
 
 def test_db_boxes_from_bitmaps():
